move layer split partitioning into core module
